support external libs
