+++ hogvm/typescript/src/stl/regex.ts
@@ -1,28 +1,49 @@
-import { RE2JS } from 're2js'
+import { ExecOptions } from '../types'
 
-export function match(regex: string, value: string, flags: number = 0): boolean {
-    // Unlike re2's default behavior, "." matches line breaks in ClickHouse by default
-    flags = flags | RE2JS.DOTALL
-    // extract all (?[ims]) and (?-[ims]) groups and apply them to the flags
-    regex = regex.replaceAll(/\(\?(-?)([ims]+)\)/g, (flagGroup, negative, modifiers) => {
+/** Extract (?-i) strings from regex */
+function gatherModifiers(regex: string, initialModifiers: string): { regex: string; modifiers: string } {
+    let insensitive = false
+    let multiline = false
+    let dotall = true // defaults to true on clickhouse
+
+    function setModifier(modifier: string, negative: boolean): void {
+        switch (modifier) {
+            case 'i':
+                insensitive = !negative
+                break
+            case 'm':
+                multiline = !negative
+                break
+            case 's':
+                dotall = !negative
+                break
+        }
+    }
+    // set defaults
+    for (const modifier of initialModifiers.split('')) {
+        setModifier(modifier, false)
+    }
+    regex = regex.replaceAll(/\(\?(-?)([ims]+)\)/g, (_, negative, modifiers) => {
         for (const modifier of modifiers.split('')) {
-            switch (modifier) {
-                case 'i':
-                    flags = negative ? flags & ~RE2JS.CASE_INSENSITIVE : flags | RE2JS.CASE_INSENSITIVE
-                    break
-                case 'm':
-                    flags = negative ? flags & ~RE2JS.MULTILINE : flags | RE2JS.MULTILINE
-                    break
-                case 's':
-                    flags = negative ? flags & ~RE2JS.DOTALL : flags | RE2JS.DOTALL
-                    break
-            }
+            setModifier(modifier, negative === '-')
         }
         return ''
     })
-    return RE2JS.compile(regex, flags).matcher(value).find()
+    return { regex, modifiers: `${insensitive ? 'i' : ''}${multiline ? 'm' : ''}${dotall ? 's' : ''}` }
 }
 
-export function matchInsensitive(regex: string, value: string, flags: number = 0): boolean {
-    return match(regex, value, RE2JS.CASE_INSENSITIVE | flags)
+export function match(regex: string, value: string, options?: ExecOptions): boolean {
+    if (options?.external?.re2) {
+        return new options.external.re2(regex, 's').test(value)
+    }
+    const { regex: newRegex, modifiers } = gatherModifiers(regex, 's')
+    return new RegExp(newRegex, modifiers).test(value)
+}
+
+export function matchInsensitive(regex: string, value: string, options?: ExecOptions): boolean {
+    if (options?.external?.re2) {
+        return new options.external.re2(regex, 'is').test(value)
+    }
+    const { regex: newRegex, modifiers } = gatherModifiers(regex, 'is')
+    return new RegExp(newRegex, modifiers).test(value)
 }
